dom/performance/tests/test_sharedWorker_performance_user_timing.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <title>Test for worker performance timing API</title>
    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body>
    <script class="testbody" type="text/javascript">

var sw = new SharedWorker('sharedworker_performance_user_timing.js');
sw.port.onmessage = function(event) {
  if (event.data.type == 'finish') {
    SimpleTest.finish();
  } else if (event.data.type == 'status') {
    ok(event.data.status, event.data.msg);
  }
}

SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
